pasando covertura en pruebas
